pyproject.toml
[project]
name = "dispatcher"
version = "0.1.0"
description = "A2A protocol dispatcher agent connecting to SQL writer and RAG agents"
readme = "README.md"
requires-python = ">=3.9,<3.14"
dependencies = [
    "ag2[openai]",
    "streamlit",
    "requests",
    "sseclient-py",
    "python-dotenv",
    "groq>=0.24.0",
]
